backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
psycopg2-binary==2.9.7
requests==2.31.0
python-dotenv==1.0.0

# 개발용 도구 (선택사항)
# pytest==7.4.2
# pytest-flask==1.2.0